requirements.txt
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
distlib==0.3.6
et-xmlfile==1.1.0
ffmpeg-python==0.2.0
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
future==0.18.2
gTTS==2.2.4
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
more-itertools==8.14.0
numpy==1.21.6
openai==0.23.1
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
pandas-stubs==1.2.0.62
pipenv==2022.10.12
platformdirs==2.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
six==1.16.0
tokenizers==0.13.1
torch==1.12.1
tqdm==4.64.1
transformers==4.23.1
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
Werkzeug==2.2.2
whisper @ git+https://github.com/openai/whisper.git@d18e9ea5dd2ca57c697e8e55f9e654f06ede25d0
wincertstore==0.2
zipp==3.9.0
